Add limits
Update test function

The math checks to see if there is room for the increment/decrement between the MAX/MIN and what the last known good result is. If there is room, it makes the math and change. If not, it stops the loop and prints the last good value and notifies the user.

Add exception handling